+++ aura/exec/Aura/Commands/P.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE LambdaCase   #-}
+
 -- |
 -- Module    : Aura.Commands.P
 -- Copyright : (c) Colin Woodbury, 2012 - 2020
@@ -10,15 +13,20 @@
 module Aura.Commands.P
   ( exploitsFromFile
   , exploitsFromStdin
+  , audit
   ) where
 
-import Aura.Core
-import Aura.Languages
-import Aura.Pkgbuild.Security
-import Aura.Security
-import Aura.Types
-import RIO
-import RIO.ByteString (getContents)
+import           Aura.Core
+import           Aura.Languages
+import           Aura.Pkgbuild.Fetch
+import           Aura.Pkgbuild.Security
+import           Aura.Security
+import           Aura.Settings
+import           Aura.Types
+import           Control.Scheduler (Comp(..), traverseConcurrently)
+import           RIO
+import           RIO.ByteString (getContents)
+import qualified RIO.Set as S
 
 ---
 
@@ -29,6 +37,28 @@ exploitsFromFile = readFileBinary >=> findExploits . Pkgbuild
 
 exploitsFromStdin :: RIO Env ()
 exploitsFromStdin = getContents >>= findExploits . Pkgbuild
+
+-- | Analyse all locally installed AUR packages.
+audit :: RIO Env ()
+audit = do
+  ss <- asks settings
+  let !m = managerOf ss
+  ps <- liftIO foreignPackages
+  warn ss . security_13 . fromIntegral $ S.size ps
+  pbs <- catMaybes <$> liftIO (traverseConcurrently Par' (getPkgbuild m . spName) $ S.toList ps)
+  mapMaybeA f pbs >>= \case
+    [] -> notify ss security_14
+    _  -> throwM $ Failure security_12
+  where
+    f :: Pkgbuild -> RIO Env (Maybe Pkgbuild)
+    f pb = case parsedPB pb of
+      Nothing -> pure Nothing
+      Just l -> case bannedTerms l of
+        [] -> pure Nothing
+        bts -> do
+          ss <- asks settings
+          liftIO $ traverse_ (displayBannedTerms ss) bts
+          pure $ Just pb
 
 findExploits :: Pkgbuild -> RIO Env ()
 findExploits pb = case parsedPB pb of